Moving to a plainer style
I received an email from GDS saying that:

1. We were using a particular crown without permission
2. We were using their New Transport font which is only meant to be used on gov.uk

All pretty sensible stuff so we'd better change to something different.

We fell foul of 1 and 2 largely due to us using some of the gov.uk template so in this commit (which is a bit too big) I've:

* Change _config.yml to go back to using the default template
* Made some changes to the default template to give us a minimal reasonably spaced design that looks okay on various devices
* Used a google font (Oswald) for headings.
* Made a couple of small copy changes

This makes the site look tidier and quite nice in a minimal sort of way.