run 11ty via it's binary rather than npx <packagename>
we were using the convention that 11ty has in it's templates,
but 11ty does register a bin file so when we're in a node
script and that will be on the path we can just use that instead.

we need to use that because how nodejs is provisioned in our
build environment, the npx executable isn't linked automatically
at the moment.